Created the payload completion callbacks
