+++ examples/pfl/fedbabu/fedavg_cifar10_levit.yml
@@ -63,7 +63,7 @@ algorithm:
     type: fedavg_partial
 
 
-    head_modules_name:
+    completion_modules_name:
         - classifier
 
     personalization:
@@ -83,7 +83,8 @@ algorithm:
 
         logging_iteration_interval: 50
         logging_epoch_interval: 1
-
+        model_logging_epoch_interval: 2
+        
         # the ratio of clients participanting in training
         participant_clients_ratio: 0.5
